add animation functionality to squares; restructure moving platforms logic
Each moving platform square should now contain a list of links. When the platform is first at this square, it is moved to the first of these links. When it returns to this square, the game cycles through the available links to choose where to send the platform to next